Fix inconsistent user id bug
Previously the userId on context object was generated in each
component, experiment lib, and analytics lib. This introduced a bug
where user ids are inconsistent across the app and made metrics
tracking impossible. This bug fix invokes getContext function once
in the root component and passes the context down to child
components as props considering the size of project.